Reintroduce RoTxn::commit: fix database opening read-only database.
Opening of databases with Env::open_database() without previous Env in memory
with a RoTxn lead to the following error:
`Io(Os { code: 22, kind: InvalidInput, message: "Invalid argument" })`

It's due to the fact that RoTxn are not commited so the env->me_numdbs cannot be
updated and will be set to CORE_DB so 2.

It can trigger from multi-process setup, read-only opening, closing write
access, and re-opening from the same process.

Thanks to @kerollmops for the typo and review.

Co-authored-by: Clément Renault <renault.cle@gmail.com>